+++ src/main/java/net/frozenblock/wilderwild/datagen/tag/WWBiomeTagProvider.java
@@ -1996,8 +1996,7 @@ public final class WWBiomeTagProvider extends FrozenBiomeTagProvider {
 		this.getOrCreateTagBuilder(WWBiomeTags.HAS_FROZEN_PLAIN_GRASS)
 			.add(Biomes.SNOWY_PLAINS)
 			.add(Biomes.ICE_SPIKES)
-			.add(Biomes.GROVE)
-			.addOptional(WWBiomes.FROZEN_CAVES);
+			.add(Biomes.GROVE);
 
 		this.getOrCreateTagBuilder(WWBiomeTags.HAS_FROZEN_TAIGA_GRASS)
 			.addOptional(WWBiomes.SNOWY_OLD_GROWTH_PINE_TAIGA);
@@ -2011,8 +2010,7 @@ public final class WWBiomeTagProvider extends FrozenBiomeTagProvider {
 
 		this.getOrCreateTagBuilder(WWBiomeTags.HAS_FROZEN_PLAIN_TALL_GRASS)
 			.add(Biomes.SNOWY_PLAINS)
-			.add(Biomes.ICE_SPIKES)
-			.addOptional(WWBiomes.FROZEN_CAVES);
+			.add(Biomes.ICE_SPIKES);
 	}
 
 	private void generateStructureTags() {